requirements.txt
faker
mysql.connector
fastapi
uvicorn[standard]